Make user think twice before adding ROOT files into the repository
Added *.root -files to the Git ignore list. This means that by default they are
not added to the repository. This behavior can be avoided by using "force"
option of the "git add" command.

The rationale for excluding ROOT files by default is that often they are quite
big and cannot be compressed further. Adding big ROOT files into the repository
would fill up our available disk space on Github and shell accounts.