Framework and tool for getting the signal efficiency in "CSV" format
At least SVM classifier needs to be trained with many different set of
parameters. The bookkeeping of which parameter set produced which
signal efficiency could be done in e.g. spreadsheet, but that would
require quite a lot of copy-pasting numbers. In order to avoid this, a
simple "framework" for outputting the relevant results to a text file
was added. It is assumed that the classifier parameters are encoded to
the classifier name separated by underscores (e.g. SVM_Gauss_2_1).

The script (code/scripts/csv.C) produces a 2D-plot, where the axes are
two classifier parameters and the points have the corresponding result
printed. The script also supports output for one classifier parameter,
in which case the y-axis is the result.

Few notes:
  - the actual data separator is actually a space, not comma
  - the output file includes a header for TTree::ReadFile()