Quick&dirty fix in chep09tmva.cc
The new (large) signal data sample has a tree with Pythia6 in it's
name, whereas e.g. the small test signal sample the tree has Pythia8
in it's name (this comes from HipProofAnalysis and the fact that the
tree names contain the name of the dataset).

This has now been circumvented by sniffing the name of the tree from
the first file on file list (signal/bkg train/test files).